Revert "Merge pull request #8 from Build-Week-5th-wheel-airbnb/jm-aws"
